fetch

+++ js/carrito.js
@@ -1,5 +1,5 @@
-document.addEventListener("DOMContentLoaded", function() {
-   
+document.addEventListener("DOMContentLoaded", function () {
+
     // Tu código JavaScript aquí
 });
 let productosEnCarrito = localStorage.getItem("productos-en-carrito");
@@ -16,11 +16,11 @@ const botonComprar = document.querySelector("#carrito-acciones-comprar");
 
 contenedorCarritoProductos.innerHTML = "";
 
-function cargarProductosCarrito(){
+function cargarProductosCarrito() {
     // Vacía el contenido actual del contenedor
     contenedorCarritoProductos.innerHTML = "";
 
-    if(productosEnCarrito && productosEnCarrito.length > 0){
+    if (productosEnCarrito && productosEnCarrito.length > 0) {
         contenedorCarritoVacio.classList.add("disabled");
         contenedorCarritoProductos.classList.remove("disabled");
         contenedorCarritoAcciones.classList.remove("disabled");
@@ -77,6 +77,26 @@ function actualizarBotonesEliminar() {
 
 
 function eliminarDelCarrito(e) {
+    Toastify({
+        text: "Producto eliminado",
+        duration: 3000,
+        newWindow: true,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true,
+        offset: {
+            x: 70, // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: 10 // vertical axis - can be a number or a string indicating unity. eg: '2em'
+        }, // Prevents dismissing of toast on hover
+        style: {
+            background: "linear-gradient(to right, #000000, #EB5757)",
+            borderRadius: ".5rem",
+            textTransform: "uppercase",
+            fontSize: ".75rem"
+        },
+        onClick: function () { } // Callback after click
+    }).showToast();
     e.stopPropagation(); // Detiene la propagación del evento
     const idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.nombre === idBoton);
@@ -98,29 +118,54 @@ function eliminarDelCarrito(e) {
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 
-function vaciarCarrito(){
+function vaciarCarrito() {
+    Swal.fire({
+        title: '¿Estás seguro?',
+        icon: 'question',
+        html:
+            `Se van a borrar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonColor: 'red',
+        confirmButtonText:
+            'Si',
+        cancelButtonText:
+            'No',
+        cancelButtonAriaLabel: 'Thumbs down'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+    })
 
-    productosEnCarrito.length = 0;
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+
 }
 
-function actualizarTotal(){
+function actualizarTotal() {
     const totalCalculado = productosEnCarrito.reduce((acc, producto) => acc + (producto.precio * producto.cantidad), 0);
     total.innerText = `$${totalCalculado}`;
 }
-    
+
 
 botonComprar.addEventListener("click", comprarCarrito);
 
-function comprarCarrito(){
+function comprarCarrito() {
+    Swal.fire({
+        position: 'center',
+        icon: 'success',
+        title: 'Tu compra a sido exitosa',
+        showConfirmButton: false,
+        timer: 1500
+      })
 
     productosEnCarrito.length = 0;
     localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
 
     contenedorCarritoVacio.classList.add("disabled");
-        contenedorCarritoProductos.classList.add("disabled");
-        contenedorCarritoAcciones.classList.add("disabled");
-        contenedorCarritoComprado.classList.remove("disabled");
+    contenedorCarritoProductos.classList.add("disabled");
+    contenedorCarritoAcciones.classList.add("disabled");
+    contenedorCarritoComprado.classList.remove("disabled");
 }
